requirements.txt
pycryptodome==3.23.0
pyperclip==1.9.0
stdiomask==0.0.6
